refactor: InputView에 대한 역할과 책임 재정의
InputView가 입력에 대한 포괄적인 의미를 담고 있다고 생각이 들어서 RacingInputView로 클래스명을 변경하였습니다.
의미 없는 메서드 분리를 삭제하였고 Racing에 대한 입력만을 책임지는 클래스로 리팩터링 하였습니다.
