+++ README.md
@@ -42,9 +42,9 @@
 
 ## STEP 4 리팩토링 요구사항
 * [ ] 움직이는 기준에 대한 책임변경
-* [ ]  InputView에 대한 역할과 책임 재정의
+* [x]  InputView에 대한 역할과 책임 재정의
 * [ ] int변환에 대한 역할은 누가 가질 것인가?
 * [ ] RacingRecord에 대한 의존관계 최소화
 * [ ] 에러 발생 원인 구체화
-* [ ] 의미 없는 메서드 분리 삭제
+* [x] 의미 없는 메서드 분리 삭제
